fix: stricter typescript rules (#202)
* chore(core): stricter types definition

* chore(http-server): stricter types definition

* chore(core): fixed test types

* chore(http-server, fastify, express, health-checks, openapi): stricter types definition

* chore(http-server): fixed test types

* chore(core): fixed object-utils types

* chore(fastify, express, messaging, messaging-amqp): fixed test types

* chore(mongoose): stricter types definition (WIP)

* chore(openapi): .eslintrc and tsconfig.json files

* chore(health-checks): .eslintrc and tsconfig.json files

* refactor: Interceptor types and generics

* fix(http-server): AjvPluginOptions type

Also added default 'topic' value to amqp subscription option exchange type

* chore: examples tsconfig

* chore: remove some ts-config and harcoded types

* chore: covecov yml indentation

* chore: covecov yml indentation #2

+++ packages/http-server-express/src/ExpressHttpServer.ts
@@ -9,7 +9,7 @@ import {
 	RequestHandler,
 	StaticServeOptions
 } from '@davinci/http-server';
-import express, { Express, Request, Response } from 'express';
+import express, { Application, Express, Request, Response } from 'express';
 import http, { Server as HttpServer } from 'http';
 import https, { Server as HttpsServer, ServerOptions } from 'https';
 import type { App } from '@davinci/core';
@@ -34,19 +34,19 @@ export class ExpressHttpServer extends HttpServerModule<{
 	Server: Server;
 	ModuleOptions: ExpressHttpServerModuleOptions;
 }> {
-	instance: Express;
-	app: App;
+	app?: App;
+	instance?: Express;
 
 	constructor(options?: ExpressHttpServerModuleOptions) {
 		super(options ?? {});
-		if (this.moduleOptions.logger?.level) {
-			this.logger.level = this.moduleOptions.logger?.level;
+		if (this.moduleOptions?.logger?.level) {
+			this.logger.level = this.moduleOptions?.logger?.level;
 		}
 	}
 
-	async onRegister(app) {
+	async onRegister(app: App) {
 		this.app = app;
-		const level = this.moduleOptions.logger?.level ?? app.getOptions().logger?.level;
+		const level = this.moduleOptions?.logger?.level ?? app.getOptions().logger?.level;
 		if (level) {
 			this.logger.level = level;
 		}
@@ -66,6 +66,7 @@ export class ExpressHttpServer extends HttpServerModule<{
 	}
 
 	registerMiddlewares() {
+		if (!this.instance) throw new Error('instance not set, aborting');
 		const { json, urlencoded, cors: corsOptions } = this.moduleOptions?.middlewares ?? {};
 
 		this.instance.use(express.json({ ...json }));
@@ -74,16 +75,15 @@ export class ExpressHttpServer extends HttpServerModule<{
 	}
 
 	initHttpServer() {
-		const { instance } = this.moduleOptions;
+		const { instance } = this.moduleOptions ?? {};
 		if (instance && typeof instance === 'function' && !(instance as Express).listen) {
 			this.instance = (instance as Function)();
 		} else {
 			this.instance = (instance as Express) ?? express();
 		}
 
-		const isHttpsEnabled = super.moduleOptions?.https;
-		const server = isHttpsEnabled
-			? https.createServer(super.moduleOptions.https, this.getInstance())
+		const server = super.moduleOptions?.https
+			? https.createServer(super.moduleOptions?.https, this.getInstance())
 			: http.createServer(this.getInstance());
 
 		super.setHttpServer(server);
@@ -100,47 +100,70 @@ export class ExpressHttpServer extends HttpServerModule<{
 		return typeof body === 'object' ? response.json(body) : response.send(String(body));
 	}
 
-	public use: Express['use'] = (...args) => {
+	// @ts-ignore
+	public use: Application['use'] = (...args) => {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.use(...args);
 	};
 
 	public get(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.get(path, handler);
 	}
 
 	public post(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.post(path, handler);
 	}
 
 	public head(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.head(path, handler);
 	}
 
 	public delete(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.delete(path, handler);
 	}
 
 	public put(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.put(path, handler);
 	}
 
 	public patch(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.patch(path, handler);
 	}
 
 	public all(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.all(path, handler);
 	}
 
 	public options(path: string, handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.options(path, handler);
 	}
 
 	public static(path: string, options?: StaticServeOptions) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.use(express.static(path, options));
 	}
 
 	listen() {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		const port = this.moduleOptions?.port || 3000;
 		this.setHttpServer(this.instance.listen(port));
 		this.logger.info(`Server listening on port: ${port}`);
@@ -167,10 +190,14 @@ export class ExpressHttpServer extends HttpServerModule<{
 	}
 
 	public setErrorHandler(handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.use(handler);
 	}
 
 	public setNotFoundHandler(handler: RequestHandler<Request, Response>) {
+		if (!this.instance) throw new Error('instance not set, aborting');
+
 		return this.instance.use(handler);
 	}
 
@@ -219,13 +246,13 @@ export class ExpressHttpServer extends HttpServerModule<{
 	}) {
 		switch (source) {
 			case 'path':
-				return request.params[name];
+				return request.params[name as string];
 
 			case 'header':
-				return request.header(name);
+				return request.header(name as string);
 
 			case 'query':
-				return request.query[name];
+				return request.query[name as string];
 
 			case 'body':
 				return request.body;
